+++ src/loader/LoaderUI.cpp
@@ -33,7 +33,7 @@ void CLoaderUI::render_contents()
 	g_gui_widgets_i->begin_tab("loader", ImGuiTabBarFlags_NoCloseWithMiddleMouseButton | ImGuiTabBarFlags_FittingPolicyScroll);
 
 	g_gui_widgets_i->add_tab_item("main", true, { 0, 0 }, [this]() { render_tab_main(); });
-	g_gui_widgets_i->add_tab_item("console", true, { 0, 0 }, [this]() { render_tab_console(); });
+	g_gui_widgets_i->add_tab_item("console", true, { 0, 0 }, [this]() { render_tab_console(); }, { 8.0f, 2.0f });
 
 	g_gui_widgets_i->end_tab();
 }